guard service

+++ src/app/theme/components/user-menu/user-menu.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AppService } from 'src/app/app.service';
 import { Router } from '@angular/router'; 
 
+
 @Component({
   selector: 'app-user-menu',
   templateUrl: './user-menu.component.html',
@@ -11,16 +12,22 @@ export class UserMenuComponent implements OnInit {
 
   constructor(public appService:AppService,public router:Router) { }
   cnx = 'False';
+  user='';
   ngOnInit() {
     this.cnx = localStorage.getItem('isLoggedIn');
-
+    this.user = localStorage.getItem('nom');
+    
   }
 
   navigate(){
     this.router.navigate(["/login"]);
   }
   logout(){
-    localStorage.setItem('isLoggedIn','false');
-    window.location.href = '/';
+    if (window.confirm("Se déconnecter ?!")) {
+      localStorage.setItem('isLoggedIn','false');
+      localStorage.setItem('token','');
+      window.location.href = '/login';
+  }
+
   }
 }
